+++ svupdater/const.py
@@ -28,13 +28,8 @@ TURRIS_REPO_HEALTH_URL = "https://repo.turris.cz/.health"
 TURRIS_REPO_HEALTH_TIMEOUT = 10
 
 # Files used for approvals handling.
-APPROVALS_ASK_FILE = "/usr/share/updater/need_approval"
-APPROVALS_STAT_FILE = "/usr/share/updater/approvals"
-# Approvals notification message
-NOTIFY_MESSAGE_CS = "Updater žádá o autorizaci akcí. Autorizaci můžete přidělit v administračním rozhraní Foris " + \
-    "v záložce 'Updater'."
-NOTIFY_MESSAGE_EN = "Your approval is required to apply pending updates. You can grant it in the Foris " + \
-    "administrative interface in the 'Updater' menu."
+APPROVALS_ASK_FILE = pathlib.Path("/usr/share/updater/need_approval")
+APPROVALS_STAT_FILE = pathlib.Path("/usr/share/updater/approvals")
 
 # File containing l10n symbols as a list of supported ones
 L10N_FILE = "/usr/share/updater/l10n_supported"